Changed folder from DTO to DTOs, documented and reimplemented inheritance of its classes and interface.
