--- src/OnionCrafter.Base/DTO/BaseAuditableDTO.cs
@@ -1,14 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace OnionCrafter.Base.DTO
-{
-    public abstract class BaseAuditableDTO : IAuditableDTO
-    {
-        public DateTime Created { get; set; }
-        public DateTime Updated { get; set; }
-    }
-}
