--- src/OnionCrafter.Base/DTO/BaseRequestAuditableDTO.cs
@@ -1,8 +1,0 @@
-﻿using OnionCrafter.Base.Wrappers.Requests;
-
-namespace OnionCrafter.Base.DTO
-{
-    public abstract class BaseRequestAuditableDTO : BaseAuditableDTO, IRequestData
-    {
-    }
-}
