--- src/OnionCrafter.Base/DTO/BaseRequestDTO.cs
@@ -1,8 +1,0 @@
-﻿using OnionCrafter.Base.Wrappers.Requests;
-
-namespace OnionCrafter.Base.DTO
-{
-    public abstract class BaseRequestDTO : BaseDTO, IRequestData
-    {
-    }
-}
