--- src/OnionCrafter.Base/DTO/BaseResponseAuditableDTO.cs
@@ -1,8 +1,0 @@
-﻿using OnionCrafter.Base.Wrappers.Responses;
-
-namespace OnionCrafter.Base.DTO
-{
-    public abstract class BaseResponseAuditableDTO : BaseAuditableDTO, IResponseData
-    {
-    }
-}
