--- src/OnionCrafter.Base/DTO/BaseResponseDTO.cs
@@ -1,8 +1,0 @@
-﻿using OnionCrafter.Base.Wrappers.Responses;
-
-namespace OnionCrafter.Base.DTO
-{
-    public abstract class BaseResponseDTO : BaseDTO, IResponseData
-    {
-    }
-}
