--- src/OnionCrafter.Base/DTO/IAuditableDTO.cs
@@ -1,8 +1,0 @@
-﻿using OnionCrafter.Base.Common;
-
-namespace OnionCrafter.Base.DTO
-{
-    public interface IAuditableDTO : IBaseDTO, IAuditableClass
-    {
-    }
-}
