--- src/OnionCrafter.Base/DTO/IBaseDTO.cs
@@ -1,6 +1,0 @@
-﻿namespace OnionCrafter.Base.DTO
-{
-    public interface IBaseDTO
-    {
-    }
-}
