+++ src/OnionCrafter.Base/DTOs/BaseDTO.cs
@@ -1,0 +1,9 @@
+﻿namespace OnionCrafter.Base.DTOs
+{
+    /// <summary>
+    /// Abstract base class for a Data Transfer Object (DTO).
+    /// </summary>
+    public abstract class BaseDTO : IBaseDTO
+    {
+    }
+}
