+++ src/OnionCrafter.Base/DTOs/BaseRequestAuditableDTO.cs
@@ -1,0 +1,11 @@
+﻿using OnionCrafter.Base.Wrappers.Requests;
+
+namespace OnionCrafter.Base.DTOs
+{
+    /// <summary>
+    /// Abstract base class for an auditable request Data Transfer Object (DTO).
+    /// </summary>
+    public abstract class BaseRequestAuditableDTO : BaseAuditableDTO, IRequestData
+    {
+    }
+}
