+++ src/OnionCrafter.Base/DTOs/BaseRequestDTO.cs
@@ -1,0 +1,11 @@
+﻿using OnionCrafter.Base.Wrappers.Requests;
+
+namespace OnionCrafter.Base.DTOs
+{
+    /// <summary>
+    /// Abstract base class for a request Data Transfer Object (DTO).
+    /// </summary>
+    public abstract class BaseRequestDTO : BaseDTO, IRequestData
+    {
+    }
+}
