+++ src/OnionCrafter.Base/DTOs/BaseResponseDTO.cs
@@ -1,0 +1,11 @@
+﻿using OnionCrafter.Base.Wrappers.Responses;
+
+namespace OnionCrafter.Base.DTOs
+{
+    /// <summary>
+    /// Abstract base class for a Response Data Transfer Object (DTO).
+    /// </summary>
+    public abstract class BaseResponseDTO : BaseDTO, IResponseData
+    {
+    }
+}
