+++ src/OnionCrafter.Base/DTOs/IBaseDTO.cs
@@ -1,0 +1,9 @@
+﻿namespace OnionCrafter.Base.DTOs
+{
+    /// <summary>
+    /// Represents a base interface for Data Transfer Objects (DTOs).
+    /// </summary>
+    public interface IBaseDTO
+    {
+    }
+}
